[FC-31159] Generalize SymlinkAndCleanup
Move the SymlinkAndCleanup component to a separate module and add error
messages for exceptions.
