+++ src/batou_ext/git.py
@@ -1,7 +1,5 @@
-import glob
 import os
 import os.path
-import shutil
 import urllib.parse
 
 import batou.component
@@ -9,6 +7,7 @@ import batou.lib.file
 import batou.lib.git
 
 import batou_ext.ssh
+from batou_ext.file import SymlinkAndCleanup
 
 
 class GitCheckout(batou.component.Component):
@@ -108,78 +107,7 @@ class GitCheckout(batou.component.Component):
         )
 
     def symlink_and_cleanup(self):
-        return SymlinkAndCleanup(self.prepared_path)
-
-
-class SymlinkAndCleanup(batou.component.Component):
-
-    namevar = "current"
-    pattern = "prepared-*"
-
-    def configure(self):
-        self.dir = os.path.dirname(self.current)
-        self.current = os.path.basename(self.current)
-
-    @staticmethod
-    def _link(path):
-        try:
-            return os.readlink(path)
-        except OSError:
-            return None
-
-    @staticmethod
-    def remove(links, el):
-        try:
-            links.remove(el)
-        except ValueError:
-            pass
-
-    def _list_removals(self):
-        candidates = glob.glob(self.pattern)
-
-        current = self._link("current")
-        last = self._link("last")
-        if current == self.current:
-            # keep last+current
-            self.remove(candidates, current)
-            self.remove(candidates, last)
-        else:
-            # keep current + new current"
-            self.remove(candidates, current)
-            self.remove(candidates, self.current)
-
-        return candidates
-
-    def verify(self):
-        with self.chdir(self.dir):
-            if self._link("current") != self.current:
-                raise batou.UpdateNeeded()
-            if self._list_removals():
-                raise batou.UpdateNeeded()
-
-    def update(self):
-        with self.chdir(self.dir):
-            current = self._link("current")
-            if current != self.current:
-                try:
-                    os.remove("current")
-                except OSError:
-                    pass
-                try:
-                    os.remove("last")
-                except OSError:
-                    pass
-                batou.output.annotate("current -> {}".format(self.current))
-                os.symlink(self.current, "current")
-                if current:
-                    batou.output.annotate("last -> {}".format(current))
-                    os.symlink(current, "last")
-            for el in self._list_removals():
-                batou.output.annotate("Removing: {}".format(el))
-                try:
-                    shutil.rmtree(el)
-                except OSError:
-                    pass
+        return SymlinkAndCleanup(self.prepared_path, pattern="prepared-*")
 
 
 class Commit(batou.component.Component):
